# api/swagger.yaml
basePath: /
definitions:
  http.accountResponse:
    properties:
      account_type:
        type: string
      balance:
        type: integer
      created_at:
        type: string
      currency_code:
        type: string
      household_id:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
...
  http.balanceResponse:
    properties:
      account_id:
        type: string
      balance:
        type: integer
    type: object
...
  http.createAccountRequest:
    properties:
      account_type:
        type: string
      currency_code:
        type: string
      name:
        type: string
    type: object
...
  http.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  http.loginResponse:
    properties:
      expires_at:
        description: RFC 3339 UTC
        type: string
      token:
        type: string
    type: object
...
  http.updateAccountBalanceRequest:
    properties:
      balance:
        type: integer
      version:
        type: integer
    type: object
  http.updateAccountNameRequest:
    properties:
      name:
        type: string
      version:
        type: integer
    type: object
...
info:
  contact: {}
  description: Personal Financial Manager REST API — double-entry accounting for households.
  title: PFM-Go API
  version: "1.0"
paths:
...
  /api/v1/households/{household_id}/accounts:
    get:
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.accountResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List accounts for a household
      tags:
      - accounts
    post:
      consumes:
      - application/json
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      - description: Account input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.createAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.accountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create an account
      tags:
      - accounts
  /api/v1/households/{household_id}/accounts/{account_id}/balance:
    get:
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      - description: Account ID (UUID)
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.balanceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get account balance
      tags:
      - ledger
...
  /api/v1/households/{household_id}/accounts/{id}:
    delete:
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      - description: Account ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate an account
      tags:
      - accounts
    get:
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      - description: Account ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.accountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get account by ID
      tags:
      - accounts
  /api/v1/households/{household_id}/accounts/{id}/balance:
    put:
      consumes:
      - application/json
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      - description: Account ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Balance update input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.updateAccountBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.accountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update account balance
      tags:
      - accounts
  /api/v1/households/{household_id}/accounts/{id}/name:
    put:
      consumes:
      - application/json
      parameters:
      - description: Household ID (UUID)
        in: path
        name: household_id
        required: true
        type: string
      - description: Account ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Name update input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.updateAccountNameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.accountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update account name
      tags:
      - accounts
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.loginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Authenticate and obtain a token
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" (obtained from POST /auth/login).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"